add NodeInfo refactor into smaller components.
