+++ client/src/components/pages/Store/Wallet/NodeInfo/__tests__/NodeError.test.jsx
