+++ client/src/components/pages/Store/Wallet/NodeInfo/__tests__/NodeInfo.test.jsx
