+++ client/src/components/pages/Store/Wallet/NodeInfo/index.js
@@ -1,0 +1,2 @@
+export { NodeError } from "./NodeError";
+export { NodeInfo } from "./NodeInfo";
